requirement.txt
requests=2.18.4
Flask=0.12.2
prettytable-0.7.2